Split cppclock class into r specific part and general part
